Rename Front to App + move dataProtectorClient to an explicit 'externals' folder
